+++ server/t__routes_nft_test.go
@@ -57,17 +57,3 @@ func TestGetNftById_NotFoundError(t *testing.T) {
 	runtime.UnmarshalBody(&body, resp.Body, r)
 	assert.Equal(fmt.Sprintf("no nfts found with id: %s", nonexistentNftId), body.Error)
 }
-
-func TestHealthcheck(t *testing.T) {
-	assert, testServer, serverUrl, r := setup(t)
-	defer teardown(testServer)
-
-	resp, err := http.Get(fmt.Sprintf("%s/health", serverUrl))
-	assert.Nil(err)
-	assertValidJSONResponse(assert, resp)
-
-	body := HealthcheckResponse{}
-	runtime.UnmarshalBody(&body, resp.Body, r)
-	assert.Equal("gallery operational", body.Message)
-	assert.Equal("local", body.Env)
-}
